Psychtoolbox v3.0.13 BETA RELEASE "Set condition1!"

This starts the 3.0.13 series and ends all support for the 3.0.12 series.

* Drop upstream support for 32-Bit Octave on Linux for Intel processors. NeuroDebian will still provide the goods, so this is no real limitation in practice for users of Debian GNU/Linux or of Ubuntu flavors and spinoffs like Linux Mint.

* Drop support for 32-Bit Octave-4 on Windows, add support for 64-Bit Octave-4 instead.

* Drop support for Octave-3 on OSX. 64-Bit Octave-4 is available as a better replacement.

* Drop support for OSX 10.8 "Mountain Lion" and OSX 10.9 "Mavericks".

* New function UniqueCell and improvements to SortCell by Diederick Niehorster.

* Some bug fixes to XOrgConfCreator by zacklb.

* Some improvements to color functions by David Brainard.

* Improved RaspberryPi support. Psychtoolbox is pretty ready for the Pi! New mex file RPiGPIOMex to control the Pi's GPIO pins and a demo RaspberryPiGPIODemo to demonstrate basic GPIO control of the 2 status LED's on the RPi 2B.

* Improvements and bug fixes to validation code for high precision display devices like VPixx Datapixx/Viewpixx/Propixx and the Cambridge Research Systems Bits+ and Bits#.

* Various bug fixes and minor improvements in 313 files.

+++ PsychSourceGL/Source/Common/Screen/SCREENDrawTexture.c
@@ -442,7 +442,7 @@ PsychError SCREENDrawTextures(void)
     if (batchIt) {
         // Signal start of new batch with numRef drawn textures, all sourced from source and
         // drawn into windo target with filterMode:
-        PsychBatchBlitTexturesToDisplay(0, numRef, source, target, NULL, NULL, 0, filterMode, 1.0);
+        PsychBatchBlitTexturesToDisplay(0, numRef, source, target, NULL, NULL, 0, (int) filterMode, 1.0);
     }
 
     // Texture blitting loop:
@@ -600,7 +600,7 @@ PsychError SCREENDrawTextures(void)
 
     if (batchIt) {
         // Finalize batch drawing:
-        PsychBatchBlitTexturesToDisplay(1, numRef, source, target, NULL, NULL, 0, filterMode, 1.0);
+        PsychBatchBlitTexturesToDisplay(1, numRef, source, target, NULL, NULL, 0, (int) filterMode, 1.0);
     }
 
     // Mark end of drawing op. This is needed for single buffered drawing:
